SoleModels

Documentation for SoleModels.

SoleModels.AbstractActiveConditionalDataset

SoleModels.AbstractBooleanCondition

SoleModels.AbstractCondition

SoleModels.AbstractConditionalAlphabet

SoleModels.AbstractConditionalDataset

SoleModels.AbstractFeature

SoleModels.AbstractLogicalBooleanCondition

SoleModels.AbstractModel

SoleModels.Branch

SoleModels.CLabel

SoleModels.ConstantModel

SoleModels.ConstrainedModel

SoleModels.DecisionForest

SoleModels.DecisionList

SoleModels.DecisionTree

SoleModels.ExternalFWDFeature

SoleModels.FeatCondition

SoleModels.FeatMetaCondition

SoleModels.FinalModel

SoleModels.FunctionModel

SoleModels.Label

SoleModels.LogicalTruthCondition

SoleModels.MixedSymbolicModel

SoleModels.MultiFrameConditionalDataset

SoleModels.NamedFeature

SoleModels.RLabel

SoleModels.Rule

SoleModels.TestOperator

SoleModels.TrueCondition

SoleModels.UnboundedExplicitConditionalAlphabet

Base.isopen

Base.rand

SoleModels.antecedent

SoleModels.apply

SoleModels.apply_test_operator

SoleModels.balanced_weights

SoleModels.bestguess

SoleModels.check_antecedent

SoleModels.check_model_constraints

SoleModels.computefeature

SoleModels.consequent

SoleModels.default_weights

SoleModels.displaymodel

SoleModels.evaluaterule

SoleModels.feasiblemodelstype

SoleModels.featvaltype

SoleModels.formula

SoleModels.immediatesubmodels

SoleModels.info

SoleModels.isminifiable

SoleModels.issymbolic

SoleModels.listimmediaterules

SoleModels.listrules

SoleModels.minify

SoleModels.negconsequent

SoleModels.outcometype

SoleModels.outputtype

SoleModels.posconsequent

SoleModels.printmodel

SoleModels.representatives

SoleModels.rulemetrics

SoleModels.submodels

SoleModels.wrap

SoleModels.CLabel — Type

const CLabel  = Union{String,Integer}
const RLabel  = AbstractFloat
const Label   = Union{CLabel,RLabel}

Types for supervised machine learning labels (classification and regression).

SoleModels.Label — Type

const CLabel  = Union{String,Integer}
const RLabel  = AbstractFloat
const Label   = Union{CLabel,RLabel}

Types for supervised machine learning labels (classification and regression).

SoleModels.AbstractActiveConditionalDataset — Type

abstract type AbstractActiveConditionalDataset{
    W<:AbstractWorld,
    A<:AbstractCondition,
    T<:TruthValue,
    FR<:AbstractFrame{W,T},
} <: AbstractConditionalDataset{W,A,T,FR} end

Abstract type for active conditional datasets, that is, conditional datasets that can be used in machine learning algorithms (e.g., they have an alphabet, can enumerate propositions and learn formulas from).

See also AbstractConditionalDataset, AbstractCondition.

SoleModels.AbstractBooleanCondition — Type

abstract type AbstractBooleanCondition end

A boolean condition is a condition that evaluates to a boolean truth value (true/false), when checked on a logical interpretation.

See also TrueCondition, LogicalTruthCondition, check, syntaxstring.

SoleModels.AbstractCondition — Type

abstract type AbstractCondition end

Abstract type for representing conditions that can be interpreted and evaluated on worlds of instances of a conditional dataset. In logical contexts, these are wrapped into Propositions.

See also Proposition, syntaxstring, FeatMetaCondition, FeatCondition.

SoleModels.AbstractConditionalAlphabet — Type

abstract type AbstractConditionalAlphabet{C<:FeatCondition} <: AbstractAlphabet{C} end

Abstract type for alphabets of conditions.

See also FeatCondition, FeatMetaCondition, AbstractAlphabet.

SoleModels.AbstractConditionalDataset — Type

abstract type AbstractConditionalDataset{
    W<:AbstractWorld,
    A<:AbstractCondition,
    T<:TruthValue,
    FR<:AbstractFrame{W,T},
} <: AbstractInterpretationSet{AbstractKripkeStructure{W,A,T,FR}} end

Abstract type for conditional datasets, that is, symbolic learning datasets where each instance is a Kripke model where conditions (see AbstractCondition), and logical formulas with conditional letters can be checked on worlds.

See also AbstractInterpretationSet, AbstractCondition.

SoleModels.AbstractFeature — Type

abstract type AbstractFeature{U<:Real} end

Abstract type for features, representing a scalar functions that can be computed on a world.

See also featvaltype, computefeature, AbstractWorld.

SoleModels.AbstractLogicalBooleanCondition — Type

abstract type AbstractLogicalBooleanCondition <: AbstractBooleanCondition end

A boolean condition based on a formula of a given logic, that is to be checked on a logical interpretation.

See also formula, syntaxstring, check, AbstractBooleanCondition.

SoleModels.AbstractModel — Type

abstract type AbstractModel{O} end

Abstract type for mathematical models that, given an instance object (i.e., a piece of data), output an outcome of type O.

See also Rule, Branch, isopen, apply, issymbolic, info, outcometype.

SoleModels.Branch — Type

struct Branch{
    O,
    C<:AbstractBooleanCondition,
    FM<:AbstractModel
} <: ConstrainedModel{O,FM}
    antecedent::C
    posconsequent::FM
    negconsequent::FM
    info::NamedTuple
end

A Branch is one of the fundamental building blocks of symbolic modeling, and has the semantics:

IF (antecedent) THEN (consequent_1) ELSE (consequent_2) END

where the antecedent is boolean condition to be tested and the consequents are the feasible local outcomes of the block.

Note that FM refers to the Feasible Models (FM) allowed in the model's sub-tree.

See also antecedent, posconsequent, negconsequent, AbstractBooleanCondition, Rule, ConstrainedModel, AbstractModel.

SoleModels.ConstantModel — Type

struct ConstantModel{O} <: FinalModel{O}
    outcome::O
    info::NamedTuple
end

The simplest type of model is the ConstantModel; it is a FinalModel that always outputs the same outcome.

Examples

julia> SoleModels.FinalModel(2) isa SoleModels.ConstantModel

julia> SoleModels.FinalModel(sum) isa SoleModels.FunctionModel
┌ Warning: Over efficiency concerns, please consider wrappingJulia Function's into FunctionWrapper{O,Tuple{SoleModels.AbstractInterpretation}} structures,where O is their return type.
└ @ SoleModels ~/.julia/dev/SoleModels/src/models/base.jl:337
true

See also apply, FunctionModel, FinalModel.

SoleModels.ConstrainedModel — Type

An AbstractModel can wrap another AbstractModel, and use it to compute the outcome. As such, an AbstractModel can actually be the result of a composition of many models, and enclose a tree of AbstractModels (with FinalModels at the leaves). In order to typebound the Feasible Models (FM) allowed in the sub-tree, the ConstrainedModel type is introduced:

abstract type ConstrainedModel{O,FM<:AbstractModel} <: AbstractModel{O} end

For example, ConstrainedModel{String, Union{Branch{String}, ConstantModel{String}}} supertypes models that with String outcomes that make use of Branch{String} and ConstantModel{String} (essentially, a decision trees with Strings at the leaves).

See also FinalModel, AbstractModel.

SoleModels.DecisionForest — Type

A Decision Forest is a symbolic model that wraps an ensemble of models

struct DecisionForest{
    O,
    C<:AbstractBooleanCondition,
    FFM<:FinalModel
} <: ConstrainedModel{O, Union{<:Branch{<:O,<:C}, <:FFM}}
    trees::Vector{<:DecisionTree}
    info::NamedTuple
end

See also ConstrainedModel, MixedSymbolicModel, DecisionList, DecisionTree

SoleModels.DecisionList — Type

struct DecisionList{
    O,
    C<:AbstractBooleanCondition,
    FM<:AbstractModel
} <: ConstrainedModel{O,FM}
    rulebase::Vector{Rule{_O,_C,_FM} where {_O<:O,_C<:C,_FM<:FM}}
    defaultconsequent::FM
    info::NamedTuple
end

A DecisionList (or decision table, or rule-based model) is a symbolic model that has the semantics of an IF-ELSEIF-ELSE block:

IF (antecedent_1)     THEN (consequent_1)
ELSEIF (antecedent_2) THEN (consequent_2)
...
ELSEIF (antecedent_n) THEN (consequent_n)
ELSE (consequent_default) END

where the antecedents are conditions to be tested and the consequents are the feasible local outcomes of the block. Using the classical semantics, the antecedents are evaluated in order, and a consequent is returned as soon as a valid antecedent is found, or when the computation reaches the ELSE clause.

Note that FM refers to the Feasible Models (FM) allowed in the model's sub-tree.

See also Rule, ConstrainedModel, DecisionTree, AbstractModel.

SoleModels.DecisionTree — Type

A DecisionTree is a symbolic model that operates as a nested structure of IF-THEN-ELSE blocks:

IF (antecedent_1) THEN
    IF (antecedent_2) THEN
        (consequent_1)
    ELSE
        (consequent_2)
    END
ELSE
    IF (antecedent_3) THEN
        (consequent_3)
    ELSE
        (consequent_4)
    END
END

where the antecedents are conditions to be tested and the consequents are the feasible local outcomes of the block.

In practice, a DecisionTree simply wraps a constrained sub-tree of Branch and FinalModel:

struct DecisionTree{
O,
    C<:AbstractBooleanCondition,
    FFM<:FinalModel
} <: ConstrainedModel{O, Union{<:Branch{<:O,<:C}, <:FFM}}
    root::M where {M<:Union{FFM,Branch}}
    info::NamedTuple
end

Note that FM refers to the Feasible Models (FM) allowed in the model's sub-tree. Also note that this structure also includes an info::NamedTuple for storing additional information.

See also ConstrainedModel, MixedSymbolicModel, DecisionList.

SoleModels.ExternalFWDFeature — Type

struct ExternalFWDFeature{U} <: AbstractFeature{U}
    name::String
    fwd::Any
end

A feature encoded explicitly as (a slice of) an FWD structure (see AbstractFWD).

See also AbstractFWD, AbstractFeature.

SoleModels.FeatCondition — Type

struct FeatCondition{U,M<:FeatMetaCondition} <: AbstractCondition
    metacond::M
    a::U
end

A scalar condition comparing a computed feature value (see FeatMetaCondition) and a threshold value a. It can be evaluated on a world of an instance of a conditional dataset.

Example: $min(V1) ≥ 10$, which translates to "Within this world, the minimum of variable 1 is greater or equal than 10."

See also AbstractCondition, negation, FeatMetaCondition.

SoleModels.FeatMetaCondition — Type

struct FeatMetaCondition{F<:AbstractFeature,O<:TestOperator} <: AbstractCondition
    feature::F
    test_operator::O
end

A metacondition representing a scalar comparison method. A feature is a scalar function that can be computed on a world of an instance of a conditional dataset. A test operator is a binary mathematical relation, comparing the computed feature value and an external threshold value (see FeatCondition). A metacondition can also be used for representing the infinite set of conditions that arise with a free threshold (see UnboundedExplicitConditionalAlphabet): ${min(V1) ≥ a, a ∈ ℝ}$.

See also AbstractCondition, negation, FeatCondition.

SoleModels.FinalModel — Type

abstract type FinalModel{O} <: AbstractModel{O} end

A FinalModel is a model which outcomes do not depend on another model. An AbstractModel can generally wrap other AbstractModels. In such case, the outcome can depend on the inner models being applied on the instance object. Otherwise, the model is considered final; that is, it is a leaf of a tree of AbstractModels.

See also ConstantModel, FunctionModel, AbstractModel.

SoleModels.FunctionModel — Type

struct FunctionModel{O} <: FinalModel{O}
    f::FunctionWrapper{O}
    info::NamedTuple
end

A FunctionModel is a FinalModel that applies a native Julia Function in order to compute the outcome. Over efficiency concerns, it is mandatory to make explicit the output type O by wrapping the Function into an object of type FunctionWrapper{O}.

TODO @Michele explain functionalargs/functionalkwargs

See also ConstantModel, FunctionWrapper, FinalModel.

SoleModels.LogicalTruthCondition — Type

struct LogicalTruthCondition{F<:AbstractFormula} <: AbstractLogicalBooleanCondition
    formula::F
end

A boolean condition that, on a given logical interpretation, a logical formula evaluates to the top of the logic's algebra.

See also formula, AbstractLogicalBooleanCondition.

SoleModels.MixedSymbolicModel — Type

A MixedSymbolicModel is a symbolic model that operaters as a free nested structure of IF-THEN-ELSE and IF-ELSEIF-ELSE blocks:

IF (antecedent_1) THEN
    IF (antecedent_1)     THEN (consequent_1)
    ELSEIF (antecedent_2) THEN (consequent_2)
    ELSE (consequent_1_default) END
ELSE
    IF (antecedent_3) THEN
        (consequent_3)
    ELSE
        (consequent_4)
    END
END

where the antecedents are conditinos and the consequents are the feasible local outcomes of the block.

In Sole.jl, this logic can implemented using ConstrainedModels such as Rules, Branchs, DecisionLists, DecisionTrees, and the be wrapped into a MixedSymbolicModel:

struct MixedSymbolicModel{O,FM<:AbstractModel} <: ConstrainedModel{O,FM}
    root::M where {M<:Union{FinalModel{<:O},ConstrainedModel{<:O,<:FM}}}
    info::NamedTuple
end

Note that FM refers to the Feasible Models (FM) allowed in the model's sub-tree.

See also ConstrainedModel, DecisionTree, DecisionList.

SoleModels.MultiFrameConditionalDataset — Type

struct MultiFrameConditionalDataset{MD<:AbstractConditionalDataset}
    modalities  :: Vector{<:MD}
end

A multi-frame conditional dataset. This structure is useful for representing multimodal datasets in logical terms.

See also AbstractConditionalDataset, minify.

SoleModels.NamedFeature — Type

struct NamedFeature{U} <: AbstractFeature{U}
    name::String
end

A feature solely identified by its name.

See also AbstractFeature.

SoleModels.RLabel — Type

const CLabel  = Union{String,Integer}
const RLabel  = AbstractFloat
const Label   = Union{CLabel,RLabel}

Types for supervised machine learning labels (classification and regression).

SoleModels.Rule — Type

struct Rule{
    O,
    C<:AbstractBooleanCondition,
    FM<:AbstractModel
} <: ConstrainedModel{O,FM}
    antecedent::C
    consequent::FM
    info::NamedTuple
end

A Rule is one of the fundamental building blocks of symbolic modeling, and has the semantics:

IF (antecedent) THEN (consequent) END

where the antecedent is a condition to be tested and the consequent is the local outcome of the block.

Note that FM refers to the Feasible Models (FM) allowed in the model's sub-tree.

See also antecedent, consequent, AbstractBooleanCondition, ConstrainedModel, AbstractModel.

SoleModels.TestOperator — Type

const TestOperator = Function

A test operator is a binary Julia Function used for comparing a feature value and a threshold. In a crisp (i.e., boolean, non-fuzzy) setting, the test operator returns a boolean value, and <, >, ≥, ≤, !=, and == are typically used.

SoleModels.TrueCondition — Type

struct TrueCondition <: AbstractLogicalBooleanCondition end

A true condition is the boolean condition that always yields true.

See also LogicalTruthCondition, AbstractLogicalBooleanCondition.

SoleModels.UnboundedExplicitConditionalAlphabet — Type

struct UnboundedExplicitConditionalAlphabet{C<:FeatCondition} <: AbstractConditionalAlphabet{C}
    metaconditions::Vector{<:FeatMetaCondition}
end

An infinite alphabet of conditions induced from a finite set of metaconditions. For example, if metaconditions = [FeatMetaCondition(UnivariateMin(1), ≥)], the alphabet represents the (infinite) set: ${min(V1) ≥ a, a ∈ ℝ}$.

See also BoundedExplicitConditionalAlphabet, FeatCondition, FeatMetaCondition, AbstractAlphabet.

Base.isopen — Method

isopen(::AbstractModel)::Bool

Return whether a model is open. An AbstractModel{O} is closed if it is always able to provide an outcome of type O. Otherwise, the model can output nothing values and is referred to as open.

Rule is an example of an open model, while Branch is an example of closed model.

See also AbstractModel.

Base.rand — Method

function Base.rand(
    rng::AbstractRNG,
    a::BoundedExplicitConditionalAlphabet;
    metaconditions::Union{Nothing,FeatMetaCondition,AbstractVector{<:FeatMetaCondition}} = nothing,
    feature::Union{Nothing,AbstractFeature,AbstractVector{<:AbstractFeature}} = nothing,
    test_operator::Union{Nothing,TestOperator,AbstractVector{<:TestOperator}} = nothing,
)::Proposition

Randomly sample a Proposition holding a FeatCondition from conditional alphabet a, such that:

if metaconditions are specified, then the set of metaconditions (feature-operator pairs)

is limited to metaconditions;

if feature is specified, then the set of metaconditions (feature-operator pairs)

is limited to those with feature;

if test_operator is specified, then the set of metaconditions (feature-operator pairs)

is limited to those with test_operator.

TODO Examples

See also BoundedExplicitConditionalAlphabet, FeatCondition, FeatMetaCondition, `AbstractAlphabet'.

SoleModels.antecedent — Method

antecedent(m::Union{Rule,Branch})::AbstractBooleanCondition

Return the antecedent of a rule/branch; that is, the condition to be evaluated upon applying the model.

See also apply, consequent, check_antecedent, Rule, Branch.

SoleModels.apply — Method

apply(
    m::AbstractModel,
    i::AbstractInterpretation;
    check_args::Tuple = (),
    check_kwargs::NamedTuple = (;),
    functional_args::Tuple = (),
    functional_kwargs::NamedTuple = (;),
    kwargs...
)::outputtype(m)

apply(
    m::AbstractModel,
    d::AbstractInterpretationSet;
    check_args::Tuple = (),
    check_kwargs::NamedTuple = (; use_memo = [ThreadSafeDict{SyntaxTree,WorldSet{worldtype(d)}}() for i in 1:ninstances(d)]),
    functional_args::Tuple = (),
    functional_kwargs::NamedTuple = (;),
    kwargs...
)::AbstractVector{<:outputtype(m)}

Return the output prediction of the model on an instance, or on each instance of a dataset. The predictions can be nothing if the model is open.

check_args and check_kwargs can influence check's behavior at the time of its computation (see check)

functional_args and functional_kwargs can influence FunctionModel's behavior when the corresponding function is applied to AbstractInterpretation (see FunctionModel, AbstractInterpretation)

See also isopen, outcometype, outputtype, AbstractModel, AbstractInterpretation, AbstractInterpretationSet.

SoleModels.apply_test_operator — Method

Apply a test operator by simply passing the feature value and threshold to the (binary) test operator function.

SoleModels.balanced_weights — Method

default_weights(Y::AbstractVector{L}) where {L<:CLabel}::AbstractVector{<:Number}

Return a class-rebalancing weight vector, given a label vector Y.

SoleModels.bestguess — Function

bestguess(
    labels::AbstractVector{<:Label},
    weights::Union{Nothing,AbstractVector} = nothing;
    suppress_parity_warning = false,
)

Return the best guess for a set of labels; that is, the label that best approximates the labels provided. For classification labels, this function returns the majority class; for regression labels, the average value. If no labels are provided, nothing is returned. The computation can be weighted.

See also CLabel, RLabel, Label.

SoleModels.check_antecedent — Method

function check_antecedent(
    m::Union{Rule,Branch},
    args...;
    kwargs...
)
    check(antecedent(m), id, args...; kwargs...)
end

Simply checks the antecedent of a rule on an instance or dataset.

See also antecedent, Rule, Branch.

SoleModels.check_model_constraints — Function

This function is used when constructing ConstrainedModels to check that the inner models satisfy the desired type constraints.

See also ConstrainedModel, Rule, Branch.

SoleModels.computefeature — Method

computefeature(f::AbstractFeature{U}, channel; kwargs...)::U where {U}

Compute a feature on a channel of an instance.

See also AbstractFeature.

SoleModels.consequent — Method

consequent(m::Rule)::AbstractModel

Return the consequent of a rule.

See also antecedent, Rule.

SoleModels.default_weights — Method

default_weights(n::Integer)::AbstractVector{<:Number}

Return a default weight vector of n values.

SoleModels.displaymodel — Method

printmodel(io::IO, m::AbstractModel; kwargs...)
displaymodel(m::AbstractModel; kwargs...)

prints or returns a string representation of model m.

Arguments

header::Bool = true: when set to true, a header is printed, displaying

the info structure for m;

show_subtree_info::Bool = false: when set to true, the header is printed for

models in the sub-tree of m;

max_depth::Union{Nothing,Int} = nothing: when it is an Int, models in the sub-tree

with a depth higher than max_depth are ellipsed with "...";

syntaxstring_kwargs::NamedTuple = (;): kwargs to be passed to syntaxstring for

formatting logical formulas.

See also SoleLogics.syntaxstring, AbstractModel.

SoleModels.evaluaterule — Method

evaluaterule(
    r::Rule{O},
    X::AbstractInterpretationSet,
    Y::AbstractVector{L}
) where {O,L<:Label}

Evaluate the rule on a labelled dataset, and return a NamedTuple consisting of:

antsat::Vector{Bool}: satsfaction of the antecedent for each instance in the dataset;

ys::Vector{Union{Nothing,O}}: rule prediction. For each instance in X:

consequent(rule) if the antecedent is satisfied,

nothing otherwise.

See also Rule, AbstractInterpretationSet, Label, check_antecedent.

SoleModels.feasiblemodelstype — Method

feasiblemodelstype(m::AbstractModel)

Return a Union of the Feasible Models (FM) allowed in the sub-tree of any AbstractModel. Note that for a ConstrainedModel{O,FM<:AbstractModel}, it simply returns FM.

See also ConstrainedModel.

SoleModels.featvaltype — Method

featvaltype(::Type{<:AbstractFeature{U}}) where {U} = U
featvaltype(::AbstractFeature{U}) where {U} = U

Return the type returned by the feature.

See also AbstractWorld.

SoleModels.formula — Method

formula(c::AbstractLogicalBooleanCondition)::AbstractFormula

Return the logical formula (see SoleLogics package) of a given logical boolean condition.

See also syntaxstring, AbstractLogicalBooleanCondition.

SoleModels.immediatesubmodels — Method

immediatesubmodels(m::AbstractModel)

Return the list of immediate child models. Note: if the model is final, then the returned list will be empty.

Examples

julia> using SoleLogics

julia> branch = Branch(SoleLogics.parseformula("p∧q∨r"), "YES", "NO");

julia> immediatesubmodels(branch)
2-element Vector{SoleModels.ConstantModel{String}}:
 SoleModels.ConstantModel{String}
YES

 SoleModels.ConstantModel{String}
NO

julia> branch2 = Branch(SoleLogics.parseformula("s→p"), branch, 42);

julia> printmodel.(immediatesubmodels(branch2));
Branch
┐ p ∧ (q ∨ r)
├ ✔ YES
└ ✘ NO

ConstantModel
42

See also submodels, FinalModel, AbstractModel.

SoleModels.info — Method

info(m::AbstractModel)::NamedTuple = m.info

Return the info structure for model m; this structure is used for storing additional information that does not affect the model's behavior. This structure can hold, for example, information about the model's statistical performance during the learning phase.

SoleModels.isminifiable — Method

isminifiable(::Any)::Bool

Return whether minification can be applied on a dataset structure. See also minify.

SoleModels.issymbolic — Method

issymbolic(::AbstractModel)::Bool

Return whether a model is symbolic or not. A model is said to be symbolic when its application relies on checking formulas of a certain logical language (see SoleLogics package) on the instance. Symbolic models provide a form of transparent and interpretable modeling.

Instead, a model is said to be functional when it encodes an algebraic mathematical function (e.g., a neural network). TODO explain listrules/cascade/rules A symbolic model is one where the computation has a rule-base structure.

See also apply, listrules, AbstractModel.

SoleModels.listimmediaterules — Method

listimmediaterules(m::AbstractModel{O} where {O})::Rule{<:O}

List the immediate rules equivalent to a symbolic model.

See also listrules, issymbolic, AbstractModel.

SoleModels.listrules — Method

listrules(m::AbstractModel; force_syntaxtree::Bool = false)::Vector{<:Rule}

Return a list of rules capturing the knowledge enclosed in symbolic model. The behavior of a symbolic model can be extracted and represented as a set of mutually exclusive (and jointly exaustive, if the model is closed) rules, which can be useful for many purposes.

The keyword argument force_syntaxtree, when set to true, causes the logical antecedents in the returned rules to be represented as SyntaxTrees, as opposed to other syntax structure (e.g., LeftmostConjunctiveForm).

Examples

TODO @Michi questi esempi non sono chiari: cosa è r2_string?

@test listrules(r2_string) isa Vector{<:Rule}
julia> print(join(displaymodel.(listrules(rule); header = false)))
┐¬(r)
└ ✔ YES

julia> print(join(displaymodel.(listrules(decision_list); header = false)))
┐(r ∧ s) ∧ t
└ ✔ YES
┐¬(r)
└ ✔ YES
┐⊤
└ ✔ YES

@test listrules(rcmodel) isa Vector{<:Rule}
julia> print(join(displaymodel.(listrules(rule_cascade); header = false)))
┐(p ∧ (q ∨ r)) ∧ ((p ∧ (q ∨ r)) ∧ (p ∧ (q ∨ r)))
└ ✔ 1

julia> print(join(displaymodel.(listrules(branch); header = false)))
┐r ∧ s
└ ✔ YES
┐r ∧ (¬(s))
└ ✔ NO
┐(¬(r)) ∧ (t ∧ q)
└ ✔ YES
┐(¬(r)) ∧ (t ∧ (¬(q)))
└ ✔ NO
┐(¬(r)) ∧ (¬(t))
└ ✔ YES

julia> print(join(displaymodel.(listrules(decision_tree); header = false)))
┐r ∧ s
└ ✔ YES
┐r ∧ (¬(s))
└ ✔ NO
┐(¬(r)) ∧ (t ∧ q)
└ ✔ YES
┐(¬(r)) ∧ (t ∧ (¬(q)))
└ ✔ NO
┐(¬(r)) ∧ (¬(t))
└ ✔ YES

julia> print(join(displaymodel.(listrules(mixed_symbolic_model); header = false)))
┐q
└ ✔ 2
┐¬(q)
└ ✔ 1.5

See also listimmediaterules, issymbolic, FinalModel, AbstractModel.

SoleModels.minify — Method

minify(dataset::D1)::Tuple{D2,Function} where {D1,D2}

Return a minified version of a dataset, as well as a backmap for reverting to the original dataset. Dataset minification remaps each scalar values in the dataset to a new value such that the overall order of the values is preserved; the output dataset is smaller in size, since it relies on values of type UInt8, UInt16, UInt32, etc.

See also isminifiable.

SoleModels.negconsequent — Method

negconsequent(m::Branch)::AbstractModel

Return the negative consequent of a branch; that is, the model to be applied if the antecedent evaluates to false.

See also antecedent, Branch.

SoleModels.outcometype — Method

outcometype(::Type{<:AbstractModel{O}}) where {O} = O
outcometype(m::AbstractModel) = outcometype(typeof(m))

Return the outcome type of a model (type).

See also AbstractModel.

SoleModels.outputtype — Method

outputtype(m::AbstractModel)

Return a supertype for the outputs obtained when applying a model. The result depends on whether the model is open or closed:

outputtype(M::AbstractModel{O}) = isopen(M) ? Union{Nothing,O} : O

Note that if the model is closed, then outputtype(m) is equal to outcometype(m).

See also isopen, apply, outcometype, AbstractModel.

SoleModels.posconsequent — Method

posconsequent(m::Branch)::AbstractModel

Return the positive consequent of a branch; that is, the model to be applied if the antecedent evaluates to true.

See also antecedent, Branch.

SoleModels.printmodel — Method

printmodel(io::IO, m::AbstractModel; kwargs...)
displaymodel(m::AbstractModel; kwargs...)

prints or returns a string representation of model m.

Arguments

header::Bool = true: when set to true, a header is printed, displaying

the info structure for m;

show_subtree_info::Bool = false: when set to true, the header is printed for

models in the sub-tree of m;

max_depth::Union{Nothing,Int} = nothing: when it is an Int, models in the sub-tree

with a depth higher than max_depth are ellipsed with "...";

syntaxstring_kwargs::NamedTuple = (;): kwargs to be passed to syntaxstring for

formatting logical formulas.

See also SoleLogics.syntaxstring, AbstractModel.

SoleModels.representatives — Method

function representatives(
    fr::AbstractMultiModalFrame{W},
    S::W,
    ::AbstractRelation,
    ::FeatMetaCondition
) where {W<:AbstractWorld}

Return an iterator to the (few) representative accessible worlds that are really necessary, upon collation, for computing and propagating truth values through existential modal operators.

This allows for some optimizations when model checking specific conditional formulas. For example, it turns out that when you need to test a formula "⟨L⟩ (MyFeature ≥ 10)" on a world w, instead of computing "MyFeature" on all worlds and then maximizing, computing it on a single world is enough to decide the truth. A few cases arise depending on the relation, the feature and the test operator (or, better, its aggregator).

SoleModels.rulemetrics — Method

rulemetrics(
    r::Rule,
    X::AbstractInterpretationSet,
    Y::AbstractVector{<:Label}
)

Calculate metrics for a rule with respect to a labelled dataset and returns a NamedTuple consisting of:

support: number of instances satisfying the antecedent of the rule divided by the total number of instances;

error:

For classification problems: number of instances that were not classified

correctly divided by the total number of instances;

For regression problems: mean squared error;

length: number of propositions in the rule's antecedent.

See also Rule, AbstractInterpretationSet, Label, evaluaterule, ninstances, outcometype, consequent.

SoleModels.submodels — Method

submodels(m::AbstractModel)

Enumerate all submodels in the sub-tree. This function is the transitive closure of immediatesubmodels; in fact, the returned list includes the immediate submodels (immediatesubmodels(m)), but also their immediate submodels, and so on.

Examples

julia> using SoleLogics

julia> branch = Branch(SoleLogics.parseformula("p∧q∨r"), "YES", "NO");

julia> submodels(branch)
2-element Vector{SoleModels.ConstantModel{String}}:
 ConstantModel
YES

 ConstantModel
NO

julia> branch2 = Branch(SoleLogics.parseformula("s→p"), branch, 42);

julia> printmodel.(submodels(branch2));
Branch
┐ p ∧ (q ∨ r)
├ ✔ YES
└ ✘ NO

ConstantModel
YES

ConstantModel
NO

ConstantModel
42

julia> submodels(branch) == immediatesubmodels(branch)
true

julia> submodels(branch2) == immediatesubmodels(branch2)
false

See also immediatesubmodels, FinalModel, AbstractModel.

SoleModels.wrap — Method

wrap(o::Any)::AbstractModel

This function wraps anything into an AbstractModel. The default behavior is the following:

when called on an AbstractModel, the model is

simply returned (no wrapping is performed);

Functions and FunctionWrappers are wrapped into a FunctionModel;

every other object is wrapped into a ConstantModel.

See also ConstantModel, FunctionModel, ConstrainedModel, FinalModel.